fixing reblog bugs
